Move BatchManager to Search
