+++ src/Search/BatchManager.hh
